factor float32.rs into platform sub-modules
